+++ R/aggregate.R
@@ -8,9 +8,7 @@
 ##' method is defined by function `FUN`.
 ##'
 ##' Adjacency matrices are an elegant way to explicitly encode for
-##' shared peptides (see example below) during aggregation. Note
-##' however that missing values can't generally be ignore using `na.rm
-##' = TRUE`, as with vector-based aggregation (see examples below).
+##' shared peptides (see example below) during aggregation.
 ##'
 ##' @section Vector-based aggregation functions:
 ##'
@@ -43,6 +41,14 @@
 ##'    peptide intensities. Shared peptides are re-used multiple
 ##'    times.
 ##'
+##' @section Handling missing values:
+##' 
+##' By default, missing values in the quantitative data will propagate
+##' to the aggregated data. You can provide `na.rm = TRUE` to most 
+##' functions listed above to ignore missing values, except for 
+##' `robustSummary()` where you should supply `na.action = na.omit` 
+##' (see `?MASS::rlm`). 
+##' 
 ##' @family Quantitative feature aggregation
 ##'
 ##' @name aggregate
@@ -100,22 +106,17 @@
 ##'                             c("S1", "S2")))
 ##' x
 ##'
-##' ## simply use na.rm = TRUE to ignore missng values
+##' ## simply use na.rm = TRUE to ignore missing values
 ##' ## during the aggregation
 ##'
 ##' (k <- LETTERS[c(1, 1, 2)])
 ##' aggregate_by_vector(x, k, colSums)
 ##' aggregate_by_vector(x, k, colSums, na.rm = TRUE)
 ##'
-##'
-##' ## NAs are propagated during the matrix
-##' ## multiplication in the aggregation functon
 ##' (adj <- matrix(c(1, 1, 0, 0, 0, 1), ncol = 2,
 ##'                dimnames = list(paste0("Pep", 1:3),
-##'                                c("A", "B"))))
-##'
-##' aggregate_by_matrix(x, adj, colSumsMat)
-##' ## not implemented
-##' try(aggregate_by_matrix(x, adj, colSumsMat, na.rm = TRUE))
-##' colSumsMat
+##'                            c("A", "B"))))
+##' aggregate_by_matrix(x, adj, colSumsMat, na.rm = FALSE)
+##' aggregate_by_matrix(x, adj, colSumsMat, na.rm = TRUE)
+##' 
 NULL