+++ R/range.R
@@ -12,18 +12,11 @@
 #' @return `logical` vector of length `length(x)`.
 #' @aliases between
 #' @family helper functions for developers
+#' @useDynLib MsCoreUtils, .registration = TRUE
 #' @export
 #' @examples
 #' between(1:4, 2:3)
-between <- function(x, range) {
-    if (!is.numeric(x))
-        stop("'x' has to be numeric.")
-    if (!is.numeric(range) || length(range) != 2L)
-        stop("'range' has to be a numeric of length 2.")
-    if (range[1L] > range[2L])
-        range[2L:1L] <- range[1L:2L]
-    range[1L] <= x & x <= range[2L]
-}
+between <- function(x, range).Call(C_between, x, range)
 
 #' @rdname range
 #'