+++ tests/testthat/test_binning.R
@@ -43,6 +43,15 @@ test_that("bin works", {
     res <- bin(vals, xs, size = 3, FUN = sum)
     ## The largest bin should contain all values larger than max(brks)
     expect_equal(res$x[length(res$x)], sum(vals[xs >= max(brks)]))
+    ## without reporting mid points.
+    res_2 <- bin(vals, xs, size = 3, FUN = sum, returnMids = FALSE)
+    expect_true(is.numeric(res_2))
+    expect_equal(res$x, res_2)
+
+    ## Ensure y is ordered (if .check = TRUE), throw error if its not
+    ## Issue #108 https://github.com/rformassspectrometry/MsCoreUtils/issues/108
+    expect_error(bin(1:5, 5:1))
+    expect_no_error(bin(1:5, 5:1, .check = FALSE))
 
     ## Check exceptions
     expect_error(bin(1:3, 1:5))